Revert "add auth service(register+jwt handling), add postgresql for auth service"
