--- backend/internal/controller/http/v1/handlers/auth/auth.go
@@ -1,12 +1,0 @@
-package auth
-
-import (
-	"context"
-
-	"github.com/kurochkinivan/HabitTracker/internal/entity"
-)
-
-type AuthService interface {
-	CreateUser(ctx context.Context, user entity.User) error
-	GetUser(ctx context.Context, id string) (entity.User, error)
-}
