--- backend/internal/controller/http/v1/middleware/logger.go
@@ -1,11 +1,0 @@
-package middleware
-
-import "net/http"
-
-type appHandler func(w http.ResponseWriter, r *http.Request)  
-
-func Logger(h appHandler) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		h(w, r)
-	}
-}
