--- backend/internal/entity/user.go
@@ -1,9 +1,0 @@
-package entity
-
-type User struct {
-	ID string `json:"id"`
-	Name string `json:"name"`
-	Email string `json:"email"`
-	Password string `json:"password"`
-	CreatedAt string `json:"created_at"`
-}
