--- backend/internal/usecase/repository/postgresql/tables.go
@@ -1,9 +1,0 @@
-package postgresql
-
-// The database schema to be used.
-const sceme = "public."
-
-// The names of tables in the database
-const (
-	usersTable = sceme + "users"
-)
