--- backend/pkg/jwtToken/jwtToken.go
@@ -1,2 +1,0 @@
-package jwttoken
-
